move search components to the same folder
