+++ src/app/components/search/search-input-control/search-input-control.component.html
