+++ src/app/components/search/search-input/search-input.component.theme.scss
