+++ src/app/components/search/search-results-row/search-results-row.component.html
@@ -1,4 +1,4 @@
-<div class="app-custom-search-row">
+<div class="aca-search-results-row">
     <div class="line">
         <span *ngIf="isFile" (click)="showPreview()" class="link"> {{ name }} </span>
