+++ src/app/components/search/search-results-row/search-results-row.component.scss
@@ -1,6 +1,6 @@
 @import 'mixins';
 
-.app-custom-search-row {
+.aca-search-results-row {
     @include flex-column;
 }
