+++ src/app/components/search/search-results/search-results.component.html
@@ -114,7 +114,7 @@
                                 key
                                 type="text">
                                     <ng-template let-context>
-                                        <app-custom-dl-row [context]="context"></app-custom-dl-row>
+                                        <aca-search-results-row [context]="context"></aca-search-results-row>
                                     </ng-template>
                             </data-column>
