+++ src/app/ui/custom-theme.scss
@@ -4,7 +4,7 @@
 @import '../components/sidenav/sidenav.component.theme';
 @import '../components/about/about.component.theme';
 @import '../components/generic-error/generic-error.component.theme';
-@import '../components/search-input/search-input.component.theme';
+@import '../components/search/search-input/search-input.component.theme';
 @import '../components/settings/settings.component.theme';
 @import '../components/current-user/current-user.component.theme';
 @import '../components/header/header.component.theme';
